feat(collection): make any change to a food item result in unconfirming it
Upon any change done a FoodItem, set the confirmed property of that FoodItem to false. This is because in the Fourth Pass, we can confirm an item, but then go back and make changes to it, which should result in it being unconfirmed and needing to be confirmed in Fourth Pass again.