feat(gibsonify_api): add multiple gibsons forms saving functionality
Add functionality for saving multiple Gibsons Forms in shared preferences local storage. This is done in the update saveForm method using the a new loadForms method. The loadForm method is changed to show last saved form only, as it will be removed. The shared preferences key is changed to reflect change to multiple saved collections.

BREAKING CHANGE: