fix: recipes not saving when created from collection
When a new recipe was created from the Collection process, it either couldn't be directly chosen for the recipe since only the Save button would be visible, or after changing to the Choose button, it would not save to api. This is a fix for both, although now it is not possible to set the "saved" property when creating a new recipe from collection. This is to be investigated later.