feat(collection): clear measurement unit and value after method change
Null measurement unit and value after changing the measurement method. This was initially attempted to be implemented by sending two more events when modifying the method, namely the events to null the unit and value, which yielded a very tricky to find bug related to indexing of FoodItems by indexOf(), which was (partially) fixed to indexing by indexWhere() using UUIDs. This needs to be fully fixed in the future.